requirements.txt
pymupdf
opencv-python
numpy
pytesseract
Pillow
